chore(random): improve performance
