+++ random/random_test.go
@@ -1,7 +1,9 @@
 package random
 
 import (
+	"math/rand"
 	"testing"
+	"time"
 )
 
 func TestStringWithCharset(t *testing.T) {
@@ -26,7 +28,7 @@ func TestStringWithCharset(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			if got := StringWithCharset(tt.args.length, tt.args.charset); got != tt.want {
-				t.Errorf("StringWithCharset() = %v, want %v", got, tt.want)
+				t.Errorf("stringWithCharset() = %v, want %v", got, tt.want)
 			}
 		})
 	}
@@ -55,5 +57,28 @@ func TestString(t *testing.T) {
 				t.Errorf("String() = %v, want %v", len(got), tt.want)
 			}
 		})
+	}
+}
+
+var seededRand = rand.New(
+	rand.NewSource(time.Now().UnixNano()))
+
+func stringWithCharsetOld(length int, charset Charset) string {
+	b := make([]byte, length)
+	for i := range b {
+		b[i] = charset[seededRand.Intn(len(charset))]
+	}
+	return string(b)
+}
+
+func BenchmarkRandStringOld(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		stringWithCharsetOld(16, Alphanumeric)
+	}
+}
+
+func BenchmarkRandStringNew(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		StringWithCharset(16, Alphanumeric)
 	}
 }
